requirements.txt
py-cord>=2.6,<3.0
python-dotenv>=1.0,<2.0
